// step14_voting_nextjs_ethers_chakra_hardhat/frontend/types/Election.ts
export declare namespace Election {
  export type CandidateStruct = {
    id: PromiseOrValue<BigNumberish>;
    name: PromiseOrValue<string>;
    voteCount: PromiseOrValue<BigNumberish>;
  };

  export type CandidateStructOutput = [BigNumber, string, BigNumber] & {
    id: BigNumber;
    name: string;
    voteCount: BigNumber;
  };

  export type VoterStruct = {
    voter: PromiseOrValue<string>;
    votedOn: PromiseOrValue<BigNumberish>;
  };

  export type VoterStructOutput = [string, BigNumber] & {
    voter: string;
    votedOn: BigNumber;
  };
}